changed integer to long

+++ src/main/java/com/example/db_project/repositories/BoardingPassRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardingPassRepository extends JpaRepository<BoardingPass, Integer> {
+public interface BoardingPassRepository extends JpaRepository<BoardingPass, Long> {
 }
